+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Write warning to log if `preferrredUsername` doesn't match cached value.
 - Don't rely on acct comparsion when verifying activity signature.
 - Use actor ID instead of webfinger address in logs.
+- Use actor ID as primary identifier instead of webfinger address.
 
 ### Fixed